# contracts/openapi/evidentia.openapi.json
{
  "components": {
    "schemas": {
      "CreateSchemaDraftRequest": {
        "additionalProperties": false,
        "description": "Create command containing only client-owned mutable content.\n\n@skyhook-story X51S43NTMRW5ASYSKJBF7FW845",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SchemaDraftContentRequest"
          }
        },
        "required": ["content"],
        "title": "CreateSchemaDraftRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Public error envelope that never contains internal exception details.\n\n@skyhook-implements REQ-012\n@skyhook-story N1ZNPJWFZYPV0MVB8FP137GRJP",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": ["error", "code"],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SchemaDraftPageResponse": {
        "additionalProperties": false,
        "description": "Stable cursor page of tenant-visible schema drafts.\n\n@skyhook-implements REQ-003\n@skyhook-story X51S43NTMRW5ASYSKJBF7FW845",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SchemaDraftResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items", "next_cursor"],
        "title": "SchemaDraftPageResponse",
        "type": "object"
      },
      "SchemaDraftResponse": {
        "additionalProperties": false,
        "description": "Mutable schema snapshot with concurrency and audit metadata.\n\n@skyhook-implements REQ-003\n@skyhook-story X51S43NTMRW5ASYSKJBF7FW845",
        "properties": {
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_id",
          "version",
          "revision",
          "content",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "SchemaDraftResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "description": "Opaque revocable browser session; never expose it to JavaScript.",
        "in": "cookie",
        "name": "evidentia_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Governed Evidentia API",
    "title": "Evidentia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/schemas/drafts": {
      "get": {
        "operationId": "schemas_list_drafts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request is not authorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Schema resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command conflicts with current state"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Drafts",
        "tags": ["schemas"]
      },
      "post": {
        "operationId": "schemas_create_draft",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request is not authorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Schema resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command conflicts with current state"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create Draft",
        "tags": ["schemas"]
      }
    },
    "/api/v1/schemas/drafts/{schema_id}": {
      "get": {
        "operationId": "schemas_get_draft",
        "parameters": [
          {
            "in": "path",
            "name": "schema_id",
            "required": true,
            "schema": {
              "title": "Schema Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request is not authorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Schema resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command conflicts with current state"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Draft",
        "tags": ["schemas"]
      },
      "put": {
        "operationId": "schemas_replace_draft",
        "parameters": [
          {
            "in": "path",
            "name": "schema_id",
            "required": true,
            "schema": {
              "title": "Schema Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceSchemaDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request is not authorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Schema resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command conflicts with current state"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Replace Draft",
        "tags": ["schemas"]
      }
    },
...
  },
...
}